kill PopoverTagsInput
